dynamic library bug fixes

+++ build.bat
@@ -1,23 +1,10 @@
-rem used to disable the echo of the commands in the batch file so it doesnt print the commands to the console
 @echo off
 
-REM Check for -clean argument
-if "%1"=="-clean" (
-   
-   rem remove the build directory
-    rmdir /s /q build
-    rem create the build directory
-    mkdir build
-)
+if "%1"=="-clean" rmdir /s /q build
 
-mkdir build
+if not exist build mkdir build
 
-REM Configure with CMake
 cmake -S . -B build
-
-REM Build
 cmake --build build
 
-REM Run executable
 build\WRender.exe
-
